requirements.txt
accelerate==1.6.0
addict==2.4.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiohttp-retry==2.8.3
aioice==0.10.1
aiortc==1.9.0
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
APScheduler==3.11.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
asyncua==1.1.6
attrs==25.3.0
av==12.3.0
babel==2.17.0
backoff==2.2.1
beautifulsoup4==4.13.4
black==25.1.0
bleach==6.2.0
blinker==1.9.0
boto3==1.35.60
botocore==1.35.99
Brotli @ file:///croot/brotli-split_1736182456865/work
cachetools==5.5.2
certifi @ file:///croot/certifi_1745939216646/work/certifi
cffi==1.17.1
cfgv==3.4.0
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.8
coloredlogs==15.0.1
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.3.2
cryptography==45.0.3
cycler==0.12.1
Cython==3.0.12
dash==3.0.4
dataclasses-json==0.6.7
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
distlib==0.3.9
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
exceptiongroup==1.2.2
executing==2.2.0
faiss-gpu==1.7.2
fastapi==0.110.3
fastjsonschema==2.21.1
filelock @ file:///croot/filelock_1744281381737/work
Flask==3.0.3
flatbuffers==25.2.10
fonttools==4.57.0
fqdn==1.5.1
frozenlist==1.6.0
fsspec==2025.3.2
gmpy2 @ file:///croot/gmpy2_1738085463648/work
google-crc32c==1.7.1
graphviz==0.20.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.31.2
humanfriendly==10.0
identify==2.6.10
idna @ file:///croot/idna_1714398848350/work
ifaddr==0.2.0
imageio==2.37.0
importlib_metadata==8.7.0
inference-gpu==0.50.1
iniconfig==2.1.0
ipykernel==6.29.5
ipython==8.36.0
ipywidgets==8.1.7
isoduration==20.11.0
isort==6.0.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2 @ file:///croot/jinja2_1741710844255/work
jiter==0.10.0
jmespath==1.0.1
joblib==1.5.0
json5==0.12.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
kiwisolver==1.4.8
lazy_loader==0.4
llvmlite==0.44.0
markdown-it-py==3.0.0
MarkupSafe @ file:///croot/markupsafe_1738584038848/work
marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.3
mkl-service==2.4.0
mkl_fft @ file:///io/mkl313/mkl_fft_1730824109137/work
mkl_random @ file:///io/mkl313/mkl_random_1730823916628/work
ml_dtypes==0.5.1
mpmath @ file:///croot/mpmath_1690848262763/work
multidict==6.4.4
mypy_extensions==1.1.0
narwhals==1.39.1
natsort==8.4.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx @ file:///croot/networkx_1737039604450/work
nodeenv==1.9.1
notebook_shim==0.2.4
numba==0.61.2
numpy==1.26.4
nvidia-ml-py==12.575.51
onnx==1.17.0
onnxruntime==1.22.0
onnxruntime-gpu==1.21.1
onnxscript==0.2.5
onnxsim==0.4.36
onnxslim==0.1.56
open3d==0.19.0
openai==1.82.0
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
overrides==7.7.0
packaging==24.2
paho-mqtt==1.6.1
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow @ file:///croot/pillow_1744613067434/work
platformdirs==4.3.8
plotly==6.1.0
pluggy==1.6.0
pre_commit==4.2.0
prometheus-fastapi-instrumentator==6.0.0
prometheus_client==0.21.1
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==6.30.2
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pybase64==1.0.2
pycparser==2.22
pydantic==2.11.5
pydantic-settings==2.7.1
pydantic_core==2.33.2
pydot==2.0.0
pyee==13.0.0
Pygments==2.19.1
PyJWT==2.10.1
pylibsrtp==0.12.0
pylogix==1.0.5
pymodbus==3.8.3
pynndescent==0.5.13
pyOpenSSL==25.1.0
pyparsing==3.2.3
pyquaternion==0.9.9
PySocks @ file:///home/builder/ci_310/pysocks_1640793678128/work
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.3.0
pytz==2025.2
PyYAML @ file:///croot/pyyaml_1728657952215/work
pyzmq==26.4.0
redis==5.0.8
referencing==0.36.2
regex==2024.11.6
requests @ file:///croot/requests_1730999120400/work
requests-file==2.1.0
requests-toolbelt==1.0.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.24.0
s3transfer==0.10.4
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
Send2Trash==1.8.3
sentencepiece==0.2.0
shapely==2.0.7
shellingham==1.5.4
six==1.17.0
slack_sdk==3.33.5
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
sports @ git+https://github.com/roboflow/sports.git@42c80c06b6b65a7f89455b89fe31cdf4c38ba227
stack-data==0.6.3
starlette==0.37.2
structlog==24.4.0
supervision==0.25.1
sympy==1.13.1
terminado==0.18.1
threadpoolctl==3.6.0
tifffile==2025.3.30
tinycss2==1.4.0
tldextract==5.1.3
tokenizers==0.21.1
tomli==2.2.1
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
torchviz==0.0.3
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.51.3
triton==3.1.0
twilio==9.3.8
typer==0.12.5
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions @ file:///croot/typing_extensions_1734714854207/work
tzdata==2025.2
tzlocal==5.3.1
ultralytics==8.3.146
ultralytics-thop==2.0.14
umap-learn==0.5.7
uri-template==1.3.0
urllib3 @ file:///croot/urllib3_1737133630106/work
virtualenv==20.31.2
wait_for2==0.3.2
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.6
widgetsnbextension==4.0.14
xformers==0.0.28.post3
yarl==1.20.0
zipp==3.21.0
zxing-cpp==2.2.0
